pyproject.toml
[project]
name = "polymarket-ai-market-suggestor"
version = "0.1.0"
description = "AI agent that analyzes trends and suggests new Polymarket prediction markets in real time."
authors = [{ name = "Lorine93s", email = "xsui46941@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
  "typer==0.12.5",
  "pydantic==2.9.2",
  "requests==2.32.3",
  "pandas==2.2.3",
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "loguru==0.7.2",
  "langchain==0.3.4",
  "langchain-openai==0.2.2",
  "chromadb==0.5.5",
  "tiktoken==0.7.0",
  "vaderSentiment==3.3.2",
  "rapidfuzz==3.9.6",
  "rich==13.9.2"
]

[project.scripts]
polysuggest = "polysuggest.cli:app"

[tool.uv]
managed = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

